requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
duckduckgo_search==7.3.0
e2b==1.0.5
e2b-code-interpreter==1.0.3
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.38.0
google-genai==1.0.0
gradio==5.9.1
gradio_client==1.5.2
h11==0.14.0
helium==5.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.56.9
lxml==5.3.0
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.3
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.11.0
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.6
safehttpx==0.1.6
safetensors==0.5.0
selenium==4.28.1
semantic-version==2.10.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smolagents==1.7.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0